Remove datafusion-macros's dependency on datafusion-expr
