+++ datafusion/expr/src/lib.rs
@@ -108,15 +108,13 @@ pub use partition_evaluator::PartitionEvaluator;
 pub use sqlparser;
 pub use table_source::{TableProviderFilterPushDown, TableSource, TableType};
 pub use udaf::{
-    aggregate_doc_sections, udaf_default_display_name, udaf_default_human_display,
-    udaf_default_return_field, udaf_default_schema_name,
-    udaf_default_window_function_display_name, udaf_default_window_function_schema_name,
-    AggregateUDF, AggregateUDFImpl, ReversedUDAF, SetMonotonicity, StatisticsArgs,
+    udaf_default_display_name, udaf_default_human_display, udaf_default_return_field,
+    udaf_default_schema_name, udaf_default_window_function_display_name,
+    udaf_default_window_function_schema_name, AggregateUDF, AggregateUDFImpl,
+    ReversedUDAF, SetMonotonicity, StatisticsArgs,
 };
-pub use udf::{
-    scalar_doc_sections, ReturnFieldArgs, ScalarFunctionArgs, ScalarUDF, ScalarUDFImpl,
-};
-pub use udwf::{window_doc_sections, ReversedUDWF, WindowUDF, WindowUDFImpl};
+pub use udf::{ReturnFieldArgs, ScalarFunctionArgs, ScalarUDF, ScalarUDFImpl};
+pub use udwf::{ReversedUDWF, WindowUDF, WindowUDFImpl};
 pub use window_frame::{WindowFrame, WindowFrameBound, WindowFrameUnits};
 
 #[cfg(test)]
