+++ datafusion/expr/src/udaf.rs
@@ -1221,37 +1221,6 @@ impl AggregateUDFImpl for AliasedAggregateUDFImpl {
     }
 }
 
-// Aggregate UDF doc sections for use in public documentation
-pub mod aggregate_doc_sections {
-    use crate::DocSection;
-
-    pub fn doc_sections() -> Vec<DocSection> {
-        vec![
-            DOC_SECTION_GENERAL,
-            DOC_SECTION_STATISTICAL,
-            DOC_SECTION_APPROXIMATE,
-        ]
-    }
-
-    pub const DOC_SECTION_GENERAL: DocSection = DocSection {
-        include: true,
-        label: "General Functions",
-        description: None,
-    };
-
-    pub const DOC_SECTION_STATISTICAL: DocSection = DocSection {
-        include: true,
-        label: "Statistical Functions",
-        description: None,
-    };
-
-    pub const DOC_SECTION_APPROXIMATE: DocSection = DocSection {
-        include: true,
-        label: "Approximate Functions",
-        description: None,
-    };
-}
-
 /// Indicates whether an aggregation function is monotonic as a set
 /// function. A set function is monotonically increasing if its value
 /// increases as its argument grows (as a set). Formally, `f` is a
