+++ datafusion/functions/src/math/monotonicity.rs
@@ -18,8 +18,8 @@
 use std::sync::LazyLock;
 
 use datafusion_common::{exec_err, Result, ScalarValue};
+use datafusion_doc::scalar_doc_sections::DOC_SECTION_MATH;
 use datafusion_expr::interval_arithmetic::Interval;
-use datafusion_expr::scalar_doc_sections::DOC_SECTION_MATH;
 use datafusion_expr::sort_properties::{ExprProperties, SortProperties};
 use datafusion_expr::Documentation;
