+++ datafusion/macros/src/user_doc.rs
@@ -22,7 +22,7 @@
 #![cfg_attr(docsrs, feature(doc_auto_cfg))]
 
 extern crate proc_macro;
-use datafusion_expr::scalar_doc_sections::doc_sections_const;
+use datafusion_doc::scalar_doc_sections::doc_sections_const;
 use proc_macro::TokenStream;
 use quote::quote;
 use syn::{parse_macro_input, DeriveInput, LitStr};
